issue #124: adding support for any_instance.should_not_receive
