+++ lib/rspec/mocks/any_instance/chain.rb
@@ -72,87 +72,6 @@ module RSpec
           self
         end
       end
-
-      # @private
-      class ExpectationChain < Chain
-
-        # @private
-        def initialize(*args, &block)
-          record(:should_receive, *args, &block)
-          @expectation_fulfilled = false
-        end
-
-        # @private
-        def expectation_fulfilled?
-          @expectation_fulfilled || constrained_to_any_of?(:never, :any_number_of_times)
-        end
-
-        private
-
-        def verify_invocation_order(rspec_method_name, *args, &block)
-        end
-
-        def invocation_order
-          @invocation_order ||= {
-            :should_receive => [nil],
-            :with => [:should_receive],
-            :and_return => [:with, :should_receive],
-            :and_raise => [:with, :should_receive]
-          }
-        end
-      end
-
-      # @private
-      class StubChain < Chain
-
-        # @private
-        def initialize(*args, &block)
-          record(:stub, *args, &block)
-        end
-
-        # @private
-        def expectation_fulfilled?
-          true
-        end
-
-        private
-
-        def invocation_order
-          @invocation_order ||= {
-            :stub => [nil],
-            :with => [:stub],
-            :and_return => [:with, :stub],
-            :and_raise => [:with, :stub],
-            :and_yield => [:with, :stub]
-          }
-        end
-
-        def verify_invocation_order(rspec_method_name, *args, &block)
-          unless invocation_order[rspec_method_name].include?(last_message)
-            raise(NoMethodError, "Undefined method #{rspec_method_name}")
-          end
-        end
-      end
-
-      # @private
-      class StubChainChain < StubChain
-
-        # @private
-        def initialize(*args, &block)
-          record(:stub_chain, *args, &block)
-        end
-
-        private
-
-        def invocation_order
-          @invocation_order ||= {
-            :stub_chain => [nil],
-            :and_return => [:stub_chain],
-            :and_raise => [:stub_chain],
-            :and_yield => [:stub_chain]
-          }
-        end
-      end
     end
   end
 end
